requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
black==23.1.0
charset-normalizer==3.0.1
click==8.1.3
discord.py==2.2.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
python-decouple==3.7
tomli==2.0.1
yarl==1.8.2
